refactored code to handle requeuing functionality in the case of blocked IP

+++ scripts/data/data_transfer/multi_download_to_s3.py
@@ -5,20 +5,21 @@ from open_whisper.preprocess import (
 import multiprocessing
 from tqdm import tqdm
 from itertools import repeat
-from typing import List
+from typing import List, Optional
 import shutil
 import tarfile
 import subprocess
 from moto3.queue_manager import QueueManager
 from moto3.s3_manager import S3Manager
 import time
+import os
 
 qm = QueueManager("whisper-downloading")
 s3m = S3Manager("mattd-public/whisper")
 
 def download(
     video_id: str, output_dir: str, audio_ext: str, lang_code: str, sub_format: str
-) -> None:
+) -> Optional[None]:
     """Download audio and transcript files for a given video
 
     Args:
@@ -28,18 +29,22 @@ def download(
         lang_code: Language code of the transcript
         sub_format: Format of the transcript
     """
-    download_audio(video_id=video_id, output_dir=output_dir, ext=audio_ext)
-    download_transcript(
+    a_status = download_audio(video_id=video_id, output_dir=output_dir, ext=audio_ext)
+    t_status = download_transcript(
         video_id=video_id,
         lang_code=lang_code,
         output_dir=output_dir,
         sub_format=sub_format,
     )
+    if a_status == "requeue" or t_status == "requeue":
+        return "requeue" 
+    return None
 
 
 def parallel_download(args) -> None:
     """Parallel download of audio and transcript files"""
-    download(*args)
+    status = download(*args)
+    return status
 
 
 def compress_dir(source_dir, tar_file):
@@ -47,7 +52,7 @@ def compress_dir(source_dir, tar_file):
         tar.add(source_dir, arcname=".")
 
 
-def download_to_s3(id_lang: List[List], group_id: str) -> None:
+def download_to_s3(id_lang: List[List], group_id: str) -> Optional[None]:
     """Download transcript and audio files
 
     Download transcript and audio files from a list of video IDs and language codes.
@@ -78,6 +83,9 @@ def download_to_s3(id_lang: List[List], group_id: str) -> None:
             )
         )
 
+    if "requeue" in out:
+        return "requeue"
+
     # compress files
     compress_dir(group_id, f"{group_id}.tar.gz")
 
@@ -90,23 +98,37 @@ def main():
             message, item = qm.get_next(visibility_timeout=(60*60)) # 1 hour visibility timeout
 
             id_lang = item["id_lang"]
+            print(f"{id_lang=}")
             group_id = item["group_id"]
+            print(f"{group_id=}")
 
-            download_to_s3(id_lang=id_lang, group_id=group_id)
-            
+            print("Downloading audio and transcript files")
+            status = download_to_s3(id_lang=id_lang, group_id=group_id)
+
             # upload metadata data about unavailable videos
-            s3m.upload_file(f"metadata/{group_id}/unavailable_videos.txt", f"metadata/{group_id}/unavailable_videos.txt")
+            if os.path.exists(f"metadata/{group_id}/unavailable_videos.txt"):
+                print("Uploading metadata (unavailable videos) files to S3")
+                s3m.upload_file(f"metadata/{group_id}/unavailable_videos.txt", f"metadata/{group_id}/unavailable_videos.txt")
+            if os.path.exists(f"metadata/{group_id}/blocked_ip.txt"):
+                print("Uploading metadata (videos affected by blocked IP) files to S3")
+                s3m.upload_file(f"metadata/{group_id}/blocked_ip.txt", f"metadata/{group_id}/blocked_ip.txt")
 
             qm.delete(message)
 
-            command = [
-                "python",
-                "scripts/data/data_transfer/upload_to_s3.py",
-                f"--group_id={group_id}",
-            ]
+            if status == "requeue":
+                print(f"Requeuing item {group_id}")
+                qm.upload([item])
+            else:
+                print("Uploading files to S3")
+                command = [
+                    "python",
+                    "scripts/data/data_transfer/upload_to_s3.py",
+                    f"--group_id={group_id}",
+                ]
 
-            proc = subprocess.Popen(command)
-            stdout, stderr = proc.communicate()
+                proc = subprocess.Popen(command)
+                stdout, stderr = proc.communicate()
+
         except IndexError: # no more items in queue to process
             break
